C API: Allow negative values for the range edges of the model
We don't want the model to be extended to negative values and this
cannot realy happen indeed. However, in some corner cases when we
calculate the range we may get nevative values. When those negative
values are casted to unsigned types the whole logic that processes
the state of the model gets corrupted, resulting into segmentation
faults.

+++ src/libkshark-model.c
@@ -153,10 +153,10 @@ static void ksmodel_set_in_range_bining(struct kshark_trace_histo *histo,
 		 * interval of the dataset.
 		 */
 		last = histo->data[histo->data_size - 1];
-		if (histo->min < histo->data[0]->ts) {
+		if (histo->min < (int64_t) histo->data[0]->ts) {
 			histo->min = histo->data[0]->ts;
 			histo->max = histo->min + corrected_range;
-		} else if (histo->max > last->ts) {
+		} else if (histo->max > (int64_t) last->ts) {
 			histo->max = last->ts;
 			histo->min = histo->max - corrected_range;
 		}
